--- a/packages/manager/.changeset/pr-10762-tech-stories-1723134446106.md
+++ b/packages/manager/.changeset/pr-10762-tech-stories-1723134446106.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tech Stories
+---
+
+Improve local Storybook performance ([#10762](https://github.com/linode/manager/pull/10762))