--- a/packages/manager/.storybook/main.ts
+++ b/packages/manager/.storybook/main.ts
@@ -31,8 +31,14 @@ const config: StorybookConfig = {
         prop.parent
           ? !/node_modules\/(?!@mui)/.test(prop.parent.fileName)
           : true,
+      // Only compile files that have stories for faster local development performance
+      include: [
+        'src/components/**/*.{ts,tsx}',
+        'src/features/Events/*',
+        'src/features/Linodes/LinodesDetail/LinodesDetailHeader/*',
+        'src/features/TopMenu/**/*.{ts,tsx}',
+      ],
     },
-
     reactDocgen: 'react-docgen-typescript',
   },
   docs: {