--- a/packages/manager/.storybook/main.ts
+++ b/packages/manager/.storybook/main.ts
@@ -32,12 +32,7 @@ const config: StorybookConfig = {
           ? !/node_modules\/(?!@mui)/.test(prop.parent.fileName)
           : true,
       // Only compile files that have stories for faster local development performance
-      include: [
-        'src/components/**/*.{ts,tsx}',
-        'src/features/Events/*',
-        'src/features/Linodes/LinodesDetail/LinodesDetailHeader/*',
-        'src/features/TopMenu/**/*.{ts,tsx}',
-      ],
+      include: ['src/components/**/*.{ts,tsx}', 'src/features/**/*.{ts,tsx}'],
     },
     reactDocgen: 'react-docgen-typescript',
   },