--- a/packages/manager/.storybook/main.ts
+++ b/packages/manager/.storybook/main.ts
@@ -1,5 +1,8 @@
 import type { StorybookConfig } from '@storybook/react-vite';
 import { mergeConfig } from 'vite';
+import { getReactDocgenTSFileGlobs } from './utils';
+
+const typeScriptFileGlobs = getReactDocgenTSFileGlobs();
 
 const config: StorybookConfig = {
   stories: [
@@ -37,7 +40,7 @@ const config: StorybookConfig = {
           ? !/node_modules\/(?!@mui)/.test(prop.parent.fileName)
           : true,
       // Only compile files that have stories for faster local development performance
-      include: ['src/components/**/*.{ts,tsx}', 'src/features/**/*.{ts,tsx}'],
+      include: typeScriptFileGlobs,
     },
     reactDocgen: 'react-docgen-typescript',
   },