--- a/packages/manager/.storybook/main.ts
+++ b/packages/manager/.storybook/main.ts
@@ -40,7 +40,9 @@ const config: StorybookConfig = {
           ? !/node_modules\/(?!@mui)/.test(prop.parent.fileName)
           : true,
       // Only compile files that have stories for faster local development performance
-      include: typeScriptFileGlobs,
+      include: /(development|test)/i.test(process.env.NODE_ENV ?? '')
+        ? typeScriptFileGlobs
+        : undefined,
     },
     reactDocgen: 'react-docgen-typescript',
   },