--- a/packages/manager/.storybook/utils.ts
+++ b/packages/manager/.storybook/utils.ts
@@ -1,0 +1,27 @@
+import globby from 'globby';
+
+const PATTERN = 'src/**/*.stories.tsx';
+
+/**
+ * Find all storybook files, then return the glob containing the parent component/feature.
+ * To be used in main.ts to tell react-docgen-typescript which files to compile.
+ *
+ * Example: src/components/Button/Button.stories.tsx -> src/components/Button/**\/*.{ts,tsx}
+ */
+export const getReactDocgenTSFileGlobs = () => {
+  const filesWithStories = globby.sync(PATTERN);
+  const files: string[] = [];
+
+  filesWithStories.forEach((file) => {
+    const execArr = /(src\/(components|features)\/[a-zA-Z]*(.|\/))/.exec(file);
+    if (execArr) {
+      const isDirectory = execArr[3] === '/';
+      const fileBlob = `${execArr[0]}${isDirectory ? '**/*.' : ''}{ts,tsx}`;
+      if (!files.includes(fileBlob)) {
+        files.push(fileBlob);
+      }
+    }
+  });
+
+  return files;
+};