--- a/packages/manager/.storybook/utils.ts
+++ b/packages/manager/.storybook/utils.ts
@@ -5,6 +5,7 @@ const PATTERN = 'src/**/*.stories.tsx';
 /**
  * Find all storybook files, then return the glob containing the parent component/feature.
  * To be used in main.ts to tell react-docgen-typescript which files to compile.
+ * https://github.com/linode/manager/pull/10762
  *
  * Example: src/components/Button/Button.stories.tsx -> src/components/Button/**\/*.{ts,tsx}
  */