## Description 📝
Performance optimizations for running Storybook locally

The main culprit is using `react-docgen-typescript`; it's parsing every file for docgen info. The package is also [no longer being maintained](https://github.com/styleguidist/react-docgen-typescript/issues/494) (the last release was in [end of 2021](https://github.com/styleguidist/react-docgen-typescript/releases)).
   - There's a faster docgen library that Storybook 8 uses, `react-docgen`, however we're missing a lot of typing by using that since it's using a shallower analysis.
   - The solution I ended up at was to keep `react-docgen-typescript`, but tighten the scope on what files are parsed for the docgen (Components and Features only). And to also disable `allowSyntheticDefaultImports` and `esModuleInterop` per the [Storybook MUI docs
](https://storybook.js.org/recipes/@mui/material#4-use-material-ui-prop-types-for-better-controls-and-docs)

The time from first paint to Intro loaded on my Intel i7 2019 MBP went from `over 1 minute` to `30s` 🎉 

## Preview 📷
| Before  | After   |
| ------- | ------- |
| <video src="https://github.com/user-attachments/assets/25294b89-ca3c-45cb-9536-9ecc55c9a437"/> | <video src="https://github.com/user-attachments/assets/677fdf2d-6708-4a6c-8d01-e1fa5003a344" /> |

## How to test 🧪
### Reproduction steps
(How to reproduce the issue, if applicable)
- On the develop branch, run Storybook locally and time the time it takes from first paint to Intro loaded

### Verification steps
(How to verify changes)
- Pull down this branch, run Storybook locally and the time from first paint to Intro loaded should be shorter than develop
- The should be no regressions in the typing and loading of stories compared to https://design.linode.com/

## As an Author I have considered 🤔

*Check all that apply*

- [x] 👀 Doing a self review
- [ ] ❔ Our [contribution guidelines](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md)
- [ ] 🤏 Splitting feature into small PRs
- [x] ➕ Adding a [changeset](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md#writing-a-changeset)
- [ ] 🧪 Providing/Improving test coverage
- [ ] 🔐 Removing all sensitive information from the code and PR description
- [ ] 🚩 Using a feature flag to protect the release
- [ ] 👣 Providing comprehensive reproduction steps
- [ ] 📑 Providing or updating our documentation
- [ ] 🕛 Scheduling a pair reviewing session
- [ ] 📱 Providing mobile support
- [ ] ♿  Providing accessibility support