@hana-linode 2024-08-15
@jaalah-akamai The main performance issue is using `react-docgen-typescript`. The performance improvement was negated in https://github.com/linode/manager/pull/10762/commits/792fc7d5332ad7865b20610c8fde989f790ecfa1 due to including all feature files.

I dug a bit deeper to see if we can automatically tell `react-docgen-typescript` which files to check without having to specify everything under the features folder and without having to manually do it. That has been addressed in https://github.com/linode/manager/pull/10762/commits/83e597067b643fc2b60a401ab1562a1c3adb7967.

I did not notice a significant difference by upgrading to 8.2, and upgrading came with a bunch of false positive logging that is a known issue so will hold off on the update until the Storybook team addresses that

![storybook 8 2 7 missing source file](https://github.com/user-attachments/assets/8ce2c379-6d47-4bc0-aaf0-dac303b60ace)